Make data associated with a specific TextEditor
This fixes two major bugs with the previous implementation:
* When Atom first started up, a request was sent off to calculate the
sizes for all active TextEditors, but the function used only worked on
the currently active TextEditor. Modify this so that all data is
associated with the proper TextEditor.

* When `import-cost` finished calculating the size of an
import, the marker would be created on the correct line... but
potentially the wrong TextEditor if the user had switched editors
between when the request was initialized and when it was resolved.