Stop trying to destroy the Marker in onDidChange
It seems that attempting to destroy the Marker from within an
onDidChange event listener on the Marker is triggering a bug, possibly
within Atom itself. As this can only trigger when the Marker would be
invalidated anyway, leave it alone for now. When `import-cost` is called
again on the file everything will be cleaned up so there will only be a
invalid Marker hanging around for a short while.

Fixes #5.