Fix blockHeight Exception
* Save Disposer

* Update package-lock.json

* Make updatePosition take only an editor

Since the Marker will always have been created when this gets called we
can rely on it for the offset into the line. Also since we need to
determine the lineHeight every time this is called just move that logic
into the function.

* Make data associated with a specific TextEditor

This fixes two major bugs with the previous implementation:
* When Atom first started up, a request was sent off to calculate the
sizes for all active TextEditors, but the function used only worked on
the currently active TextEditor. Modify this so that all data is
associated with the proper TextEditor.

* When `import-cost` finished calculating the size of an
import, the marker would be created on the correct line... but
potentially the wrong TextEditor if the user had switched editors
between when the request was initialized and when it was resolved.

* Update positions on change of active item

When Atom is first restoring a session a TextEditor can be created in a
tab that isn't currently displayed. `getLineHeightInPixels` returns
`null` in this case. Work around this issue by re-measuring the
positions when the active Pane item stops changing.

* Make decorations click through

Without this style when a user clicks where the decoration is, nothing
happens. With this change the decoration isn't interactable, so clicking
on it acts as if it wasn't there. This means for example that clicking
in the decoration will move the cursor to the line.

* Stop trying to destroy the Marker in onDidChange

It seems that attempting to destroy the Marker from within an
onDidChange event listener on the Marker is triggering a bug, possibly
within Atom itself. As this can only trigger when the Marker would be
invalidated anyway, leave it alone for now. When `import-cost` is called
again on the file everything will be cleaned up so there will only be a
invalid Marker hanging around for a short while.

Fixes #5.

* Handle done event

When `import-cost` claims to be finished calculating sizes remove the
Markers for any imports that couldn't be calculated.

* Prevent Marker Blip

* Editor Error Handling

* Fix Attribution